% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipletSynthMethods.R
\name{adjustMultuplets}
\alias{adjustMultuplets}
\title{adjustMultuplets}
\arguments{
\item{singletsFull}{data.frame The singlets variable output from the
syntheticSinglets function.}

\item{multiplets}{data.frame. The multiplets data.frame or a empty data.frame
with rows equal to the number of genes in the desired multiplets variable.}

\item{nGenes}{numeric. Number of genes in generated synthetic data.}

\item{nCellTypes}{numeric. Number of cell types in generated synthetic data.
For the frequency adjustment to work this should be an even number.}

\item{perplexity}{numeric. The perplexity parameter supplied to
spUnsupervised.}

\item{target}{numeric. The target percentage of connections for interacting
cells.}

\item{...}{additional arguments to pass on}
}
\value{
The multiplets variable.
}
\description{
This unit adjusts the multiplets so that each cell type has one prevelant
connection. Which connection a cell type prefers is decided randomly. The
strength of the preference is determined by the target argument. This works
well as long as the number of cell types in the dataset is even (and not
odd).
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
