% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{adjustSelf}
\alias{adjustSelf}
\title{adjustSelf}
\arguments{
\item{multuplets}{data.frame. The multiplets data.frame.}

\item{singletsFull}{data.frame The singlets variable output from the
syntheticSinglets function.}

\item{...}{additional arguments to pass on.}
}
\value{
The targetConnections data.frame.
}
\description{
Adjust the number of self connections. This currently dilutes the current
connections into a third of the total connections by adding twice as many
self connections.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
