% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionDataTest.R
\name{anyDataTest}
\alias{anyDataTest}
\title{anyDataTest}
\arguments{
\item{counts}{matrix. Counts data.}

\item{mulFrac}{numeric. The factions of the submitted data that should be
reserved for synthesizing multiplets.}

\item{outPath}{Path to output results to.}

\item{cores}{Number of cores to be used for spSwarm.}

\item{nMultiplets}{Number of multuplets.}

\item{nGenes}{Specifies the \code{n} argument to the \code{spTopMax}
function.}

\item{distFun}{The distance function to use during swarm optimization.}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
expressionDataTest()



}
\author{
Jason T. Serviss
}
\keyword{anyDataTest}
