% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{calculateConnections}
\alias{calculateConnections}
\title{calculateConnections}
\arguments{
\item{codedSwarm}{Output in spSwarm object.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame
}
\description{
Subtitle
}
\examples{

#use demo data
calculateConnections()



}
\author{
Jason T. Serviss
}
\keyword{calculateConnections}

