% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipletSynthMethods.R
\name{calculateFreqAndSum}
\alias{calculateFreqAndSum}
\title{calculateFreqAndSum}
\arguments{
\item{i}{integer. Current iteration.}

\item{targetConnections}{data.frame. Output from the decideConnections unit.}

\item{current}{data.frame. Output from the quantifyConnections unit.}

\item{...}{additional arguments to pass on.}
}
\value{
A list containing the current cell types, the current frequency of
   of the prefered connection, and the sum of the current number of
   connections between the cell types.
}
\description{
Calculates statistics concerning the current connection. These are used
downstream to calculate the number of connections that need to be added to
meet the target percentage.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
