% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipletSynthMethods.R
\name{calculateNumToAdd}
\alias{calculateNumToAdd}
\title{calculateNumToAdd}
\arguments{
\item{i}{integer. Current iteration.}

\item{currentFreq}{numeric. The current frequency of of the prefered
connection.}

\item{currentSum}{numeric.}

\item{targetConnections}{data.frame. Output from the decideConnections unit.}

\item{...}{additional arguments to pass on.}
}
\value{
Integer specifying the number of connections to add in order to meet
   the target percentage.
}
\description{
Calculates the number of connections that need to be added to meet the target
percentage for each connection.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
