% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypesIncludedTest.R
\name{cellTypesIncludedTest}
\alias{cellTypesIncludedTest}
\title{cellTypesIncludedTest}
\arguments{
\item{dataType}{Data type to use for test. Can be "synthetic" or "expression"}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
cellTypesIncludedTest()



}
\author{
Jason T. Serviss
}
\keyword{cellTypesIncludedTest}

