% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{checkFrequency}
\alias{checkFrequency}
\title{checkFrequency}
\arguments{
\item{i}{integer. Current iteration.}

\item{currentTypes}{data.frame. Output from the quantifyConnections unit.}

\item{multuplets}{data.frame. The multiplets data.frame.}

\item{targetConnections}{data.frame. Output from the decideConnections unit.}

\item{currentTypes}{The current cell types being examined.}

\item{...}{additional arguments to pass on.}
}
\value{
The targetConnections data.frame.
}
\description{
Performs a final check to see if the target frequencies have been met. The
current percentages are added to the targetConnections data.frame which is
subsequently checked outside of the unit.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
