% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMseq.testing-package.R
\docType{package}
\name{CIMseq.testing-package}
\alias{CIMseq.testing-package}
\alias{CIMseq.testing}
\title{Testing and analysis of the CIMseq and method.}
\description{
Description
}
\details{
\tabular{ll}{ Package: \tab CIMseq\cr Type: \tab Package\cr
Version: \tab 1.0\cr Date: \tab 2016-02-28\cr License: \tab GPL-3\cr }
}
\author{
Author: Jason T. Serviss
}
\keyword{package}
