% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamingAndOrdering.R
\name{classOrder.MGA}
\alias{classOrder.MGA}
\title{classOrder.MGA}
\arguments{
\item{type}{character; Indicates which order subset to return. Can be "total",
"si", or "c".}
}
\value{
A character vector with class order
}
\description{
Returns the desired class order for the MGA dataset.
}
\author{
Jason Serviss
}
