% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsSorted.R
\name{countsSortedTest}
\alias{countsSortedTest}
\title{countsSortedTest}
\arguments{
\item{data}{.}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
\dontrun{countsSortedTest(data = "countsSorted1.rda")}



}
\author{
Jason T. Serviss
}
