% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipletSynthMethods.R
\name{decideConnections}
\alias{decideConnections}
\title{decideConnections}
\arguments{
\item{cellTypes}{character. Provides the names of the cell types to include
in the multiplet(s). These must be present in the singlets matrix and the
names specified correctly in the colnames.}

\item{target}{numeric. The target percentage of connections for interacting
cells.}

\item{...}{additional arguments to pass on.}
}
\value{
data.frame
}
\description{
This randomly assigns connections between all cell types and returns a
data.frame to allow tracking of the target and current connection percentage
per connection.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
