% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentalDataTest.R
\name{experimentalDataTest}
\alias{experimentalDataTest}
\title{experimentalDataTest}
\arguments{
\item{outPath}{Path to output results to.}

\item{cores}{Number of cores to be used for spSwarm.}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
expDataTest()



}
\author{
Jason T. Serviss
}
\keyword{experimentalDataTest}
