% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipletSynthMethods.R
\name{generateMultiplets}
\alias{generateMultiplets}
\title{generateMultiplets}
\arguments{
\item{...}{additional arguments to pass on}
}
\value{
The multiplets variable.
}
\description{
generateMultiplets
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
