% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingNewConditions.R
\name{generateTestSummary}
\alias{generateTestSummary}
\title{Generate a summary when testing different distance functions.}
\arguments{
\item{spSwarms}{Output from testDistanceFunctions function.}

\item{metadata}{A tibble including sample metadata. Provided to
\code{setupPlate} function.}

\item{...}{additional arguments to pass on}
}
\value{
mean accuracy
}
\description{
Generates a full summary from a list of spSwarm objects that compares results
to the expected results and generates TRP, TNR, and accuracy metrics.
}
\author{
Jason T. Serviss
}
