% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipletSynthMethods.R
\name{makeMultuplet2}
\alias{makeMultuplet2}
\title{makeMultuplet2}
\arguments{
\item{nCellsInMultiplet}{numeric. Defines the number of cells to be included
in the multiplet. Currently can only be 2, 3, or 4.}

\item{cellTypes}{character. Provides the names of the cell types to include
in the multiplet(s). These must be present in the singlets matrix and the
names specified correctly in the colnames.}

\item{multiplets}{data.frame. The multiplets data.frame or a empty data.frame
with rows equal to the number of genes in the desired multiplets variable.}

\item{singlets}{data.frame The singlets variable output from the
syntheticSinglets function.}

\item{repetitions}{numeric. The number of times the connections should be
repeated or NULL if they should only be repeated once.}

\item{self}{logical. Indicates if self connections should be included.}

\item{fractions}{data.frame. A data frame indicating the fraction that each
cell type should contribute to a multiplet. For example, if the fraction
is 0.2, 0.4, 0.4 and the multiplet being synthesized is cell types A, B,
and C then A contributes 20%  of the multiplet counts whereas B and C both
contribute 40%. The data frame should contain 2 columns named
\code{cellTypes} and including the cell type names and the other,
named \code{fraction}, containing the fraction of each cell types
contribution.}

\item{...}{additional arguments to pass on}
}
\value{
The multiplets variable.
}
\description{
This unit synthesizes "all possible" multiplets combinations from the
singlets that were generated. Currently, doublets, triplets, and quadruplets
are synthesized. The unit also calls the adjustMultuplets and adjustSelf
units to adjust the connections in a manner that 1) self connections are the
most prevelant connection type (currently 2/3 connections) and 2) that each
cell type has a high percentage of connections to one other cell type.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
