% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseDataTest.R
\name{noiseDataTest}
\alias{noiseDataTest}
\title{noiseDataTest}
\arguments{
\item{dataType}{Data type to use for test. Can be "synthetic" or "expression"}

\item{cores}{Number of cores to run spSwarm on.}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
noiseDataTest()



}
\author{
Jason T. Serviss
}
\keyword{noiseDataTest}
