% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseDataTest.R
\name{noiseDataTestPlot}
\alias{noiseDataTestPlot}
\title{noiseDataTestPlot}
\arguments{
\item{cores}{Number of cores to run spSwarm on.}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
noiseDataTestPlot()



}
\author{
Jason T. Serviss
}
\keyword{noiseDataTestPlot}
