% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette}
\alias{palette}
\title{palette}
\usage{
palette(type)
}
\arguments{
\item{type}{character; Indicates which order subset to return. Can be "total",
"si", or "c".}
}
\value{
A named character vector with names corresponding to classes and 
elements corresponding to colours.
}
\description{
Figure palette
}
\author{
Jason T. Serviss
}
