% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationFunctions.R
\name{permCostPlot}
\alias{permCostPlot}
\title{permCostPlot}
\arguments{
\item{data}{Output from the tidyPermutationData function.}

\item{...}{additional arguments to pass on}
}
\value{
A tibble with the real and permuted costs and fractions.
}
\description{
A function to plot the permuted vs real costs.
}
\author{
Jason T. Serviss
}
