% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunPermutations.R
\name{permutationCosts}
\alias{permutationCosts}
\title{permutationCosts}
\arguments{
\item{permutationMeans}{Output from permutationMeans function.}

\item{spCountsMul}{An spCounts object containing multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{nPermsPerMul}{Numeric; Number of permutations per multiplet. If null
this is set to 100.}

\item{Function;}{A distance function that returns the cost.}

\item{...}{additional arguments to pass on.}
}
\value{
matrix
}
\description{
Calculates the costs for each permutation using the provided function and
returns these together with the real costs. Note that the real costs in the
spSwarm object should be calculated with the SAME distance function as
provided in the \code{distFun} argument.
}
\author{
Jason T. Serviss
}
