% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunPermutations.R
\name{permuteClass}
\alias{permuteClass}
\title{permuteClass}
\arguments{
\item{spCountsMul}{A spCounts object containing multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{nPerms}{Number of total permutations. If NULL set to 100 per multiplet.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame
}
\description{
Sets up permutations of the classes corresponding to the counts matrix
for evaluating the performance of various distance functions used with the
spSwarm method.
}
\author{
Jason T. Serviss
}
