% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunPermutations.R
\name{permuteCounts}
\alias{permuteCounts}
\title{calculateMeans}
\arguments{
\item{perms}{Output from \code{permuteClass} function.}

\item{spCountsSng}{A spCounts object containing singlets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame
}
\description{
Calculates the group means using output from the permuteClass function.
}
\author{
Jason T. Serviss
}
