% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationFunctions.R
\name{permuteGroupMeans}
\alias{permuteGroupMeans}
\title{permuteGroupMeans}
\arguments{
\item{spUnsupervised}{An spUnsupervised object.}

\item{...}{additional arguments to pass on}
}
\value{
Outputs one matrix of permuted group means with dims equal to the
origional group means matrix.
}
\description{
Permutes the matrix in the groupMeans slot of the spUnsupervised object that
is input. To be called repeatedly to generate multiple permutations and used
for evaluating the performance of the spSwarm method. Permutation is per row
and, thus, the values in each row are randomly shuffled.
}
\author{
Jason T. Serviss
}
