% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunPermutations.R
\name{permuteMeans}
\alias{permuteMeans}
\title{permuteMeans}
\arguments{
\item{spCountsMul}{A spCounts object containing multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{nPerms}{Number of total permutations. If NULL set to 100 per multiplet.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame
}
\description{
Sets up permutations of the classes for evaluating the performance
of various distance functions used with the spSwarm method. Subsequently
used with the \code{permutationMeans} function.
}
\author{
Jason T. Serviss
}
