% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationFunctions.R
\docType{methods}
\name{permuteSwarm}
\alias{permuteSwarm}
\alias{permuteSwarm,spCounts-method}
\title{permuteSwarm}
\usage{
permuteSwarm(spCountsMul, ...)

\S4method{permuteSwarm}{spCounts}(spCountsMul, spUnsupervised,
  distFun = sp.scRNAseq:::dtsnCellNum, maxiter = 10, swarmsize = 50,
  cores = 1, norm = TRUE, iter, cellNumbers, e = 0.0025, ...)
}
\arguments{
\item{spCountsMul}{An spCounts object containing multiplets.}

\item{...}{additional arguments to pass on}

\item{spUnsupervised}{An spUnsupervised object.}

\item{distFun}{The distance function used to calculate the cost. Either the
name of a custom function in the local environment or one of the included
functions, i.e. \code{distToSlice, distToSliceNorm, distToSliceTop,
distToSliceEuclid, distToSlicePearson, bic}.}

\item{maxiter}{pySwarm argument indicating maximum optimization iterations.}

\item{swarmsize}{pySwarm argument indicating the number of swarm particals.}

\item{cores}{The number of cores to be used while running spSwarm.}

\item{norm}{Logical indicating if the sum of fractions should equal 1.}

\item{iter}{The number of permutations to perform.}

\item{cellNumbers}{Tibble; Output from \code{estimateCells} function.}

\item{e}{Numeric; The epsilon value for the .complexityPenilty unit.}
}
\value{
?.
}
\description{
Write something
}
\details{
Description
}
\examples{

#use demo data


#run function


}
\author{
Jason T. Serviss
}
\keyword{permuteSwarm}
