% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypesIncludedTest.R
\name{plotCellTypesIncludedTest}
\alias{plotCellTypesIncludedTest}
\title{plotCellTypesIncludedTest}
\arguments{
\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
plotCellTypesIncludedTest()



}
\author{
Jason T. Serviss
}
\keyword{plotCellTypesIncludedTest}
