% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{plotConnectionDist}
\alias{plotConnectionDist}
\title{plotConnectionDist}
\arguments{
\item{multipletNames}{character. A character vector of multiplet adhering to
the "m.A1B1_1" convention.}

\item{...}{additional arguments to pass on.}
}
\value{
data.frame
}
\description{
Plots the results from the testing of the synthetic dataset.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
