% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFunPermutations.R
\name{plotDistPermutations}
\alias{plotDistPermutations}
\title{plotDistPermutations}
\arguments{
\item{permCosts}{An spCounts object containing multiplets.}

\item{facet}{An spUnsupervised object.}

\item{...}{additional arguments to pass on.}
}
\value{
matrix
}
\description{
Plots the results from the \code{permutationCosts} function.
}
\author{
Jason T. Serviss
}
