% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{plotExpVSObs}
\alias{plotExpVSObs}
\title{plotExpVSObs}
\arguments{
\item{spSwarm}{spSwarm. An spSwarm object.}

\item{edge.cutoff}{numeric. The minimum fraction to consider (?).}

\item{summary}{logical. Indicates if edges should be summarized or not.}

\item{...}{additional arguments to pass on.}
}
\value{
data.frame
}
\description{
plotExpVSObs
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
