% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSizeTest.R
\name{plotGroupSizeTest}
\alias{plotGroupSizeTest}
\title{plotGroupSizeTest}
\arguments{
\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
plotGroupSizeTest()



}
\author{
Jason T. Serviss
}
\keyword{plotGroupSizeTest}

