% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipletSynthMethods.R
\name{plotMultipletsCellNrs}
\alias{plotMultipletsCellNrs}
\title{plotMultipletsCellNrs}
\arguments{
\item{multipletNames}{character. A character vector of multiplet adhering to
the "m.A1B1_1" convention.}

\item{...}{additional arguments to pass on.}
}
\value{
data.frame
}
\description{
plotMultipletsCellNrs
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
