% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPDF}
\alias{plotPDF}
\title{plotPDF}
\usage{
plotPDF(plot, strip.text.size = 3.25)
}
\arguments{
\item{plot}{A ggplot}

\item{strip.text.size}{Numeric size.}
}
\value{
A ggplot.
}
\description{
Plotting theme
}
\author{
Jason T. Serviss
}
