% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRmarkdown}
\alias{plotRmarkdown}
\title{plotRmarkdown}
\usage{
plotRmarkdown(plot)
}
\arguments{
\item{plot}{A ggplot}

\item{strip.text.size}{Numeric size.}
}
\value{
A ggplot.
}
\description{
Plotting theme
}
\author{
Jason T. Serviss
}
