% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funToCheckKnownResults.R
\name{printResults}
\alias{printResults}
\title{printResults}
\arguments{
\item{data}{Output from checkResults function.}

\item{addFractions}{Logical indicating if fractions should be added to the
output.}

\item{spSwarm}{An spSwarm object. Required if \code{addFractions} is TRUE.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame
}
\description{
Takes the output from checkResults and prints it in a more verbose form.
(note: rewrite as general to take an arg for nested columns to expand)
}
\author{
Jason T. Serviss
}
