% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipletSynthMethods.R
\name{quantifyConnections}
\alias{quantifyConnections}
\title{quantifyConnections}
\arguments{
\item{multupletNames}{character. Names of the multiplets specified using the
typical LETTER NUMBER convention.}

\item{...}{additional arguments to pass on.}
}
\value{
data.frame
}
\description{
Uses the multuplet naming convention to quantify the frequency of each
connection.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
