% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamingAndOrdering.R
\name{renameClasses.MGA}
\alias{renameClasses.MGA}
\title{renameClasses.MGA}
\arguments{
\item{class}{character; Old class names.}

\item{abrev}{logical; Should tissue abbreviation be included?}
}
\value{
A character vector with new class names.
}
\description{
MGA dataset class renaming (from Seurat default) including tissue 
abbreviation.
}
\author{
Jason Serviss
}
