% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamingAndOrdering.R
\name{renameClasses.SCM}
\alias{renameClasses.SCM}
\title{renameClasses.SCM}
\arguments{
\item{class}{character; Old class names.}
}
\value{
A character vector with new class names.
}
\description{
SCM dataset class renaming (from Seurat default).
}
\author{
Jason Serviss
}
