% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funToCheckKnownResults.R
\name{setupPlate}
\alias{setupPlate}
\title{setupPlate}
\arguments{
\item{plateData}{tibble; Includes the columns \code{row, column,
multipletName, multipletComposition, connections} where \code{row} and
\code{column} includes a letter or numerical value, respectivley, indicating
the index in the plate of each multiplet, \code{multipletName} corresponds
to the name of the multiplet and matches the names in the subsequent
detected results, \code{multipletComposition} is the cell types included in
the multiplet seperated by a "-", and \code{connections} is a list for each
multiplet containing a matrix with all possible combinations of 2 of the
cell types included in the multiplet.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame
}
\description{
Sets up the CIMseqSwarm object for the \code{known} argument to the
\code{checkResults} function.
}
\author{
Jason T. Serviss
}
