% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seuratToCIMseq}
\alias{seuratToCIMseq}
\title{seuratToCIMseq}
\arguments{
\item{seurat.obj}{Seurat; A Seurat object.}

\item{singlets}{matrix; A matrix with singlet counts.}

\item{singlet.ercc}{matrix; A matrix with singlet ERCC counts.}

\item{multiplets}{matrix; A matrix with multiplet counts.}

\item{multiplets}{matrix; A matrix with multiplet ERCC counts.}
}
\value{
A list with CIMseqSinglets object as the first element and 
CIMseqMultiplets object as the second element.
}
\description{
Exttracts info from Seurat object and returns CIMseqSinglets and 
CIMseqMultiplets objects.
}
\author{
Jason T. Serviss
}
