% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationFunctions.R
\name{spSwarmPermutationData}
\alias{spSwarmPermutationData}
\title{spSwarmPermutationData}
\arguments{
\item{sObj}{An spSwarm object where the "real" data is included.}

\item{permData}{Output from the permuteSwarm function.}

\item{...}{additional arguments to pass on}
}
\value{
A list of spSwarm objects.
}
\description{
A function to put the results from the \code{permuteSwarm} function into
spSwarm objects.
}
\author{
Jason T. Serviss
}
