% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{synthesizeAndAdd}
\alias{synthesizeAndAdd}
\title{synthesizeAndAdd}
\arguments{
\item{add}{integer. The number of connections to add.}

\item{currentTypes}{data.frame. Output from the quantifyConnections unit.}

\item{multuplets}{data.frame. The multiplets data.frame or a empty data.frame
with rows equal to the number of genes in the desired multiplets variable.}

\item{singletsFull}{data.frame The singlets variable output from the
syntheticSinglets function.}

\item{...}{additional arguments to pass on.}
}
\value{
The multiplets variable with the additional multiplets added by the
   unit.
}
\description{
Synthesizes the number of connections necessary to meet the target
percentages according to the add argument. Doublets, triplets, and
quadruplets are synthesized.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
