% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{syntheticDataTest}
\alias{syntheticDataTest}
\title{syntheticDataTest}
\arguments{
\item{outPath}{Path to output results to.}

\item{cores}{Number of cores to be used for spSwarm.}

\item{n}{Number of multuplets.}

\item{ngenes}{Number of genes in generated synthetic data.}

\item{ncells}{Number of cells per cell type in generated synthetic data.}

\item{cellTypes}{Number of cell types in generated synthetic data.}

\item{distFun}{The distance function to use for swarm optimization.}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
syntheticDataTest()



}
\author{
Jason T. Serviss
}
\keyword{syntheticDataTest}
