% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticMultipletsFromCounts.R
\name{syntheticMultipletsFromCounts}
\alias{syntheticMultipletsFromCounts}
\title{syntheticMultipletsFromCounts}
\arguments{
\item{counts}{matrix; A matrix of singlet counts per million with cells as
columns and genes as rows.}

\item{classes}{character; A character vector indicating the cell types of the
cells in the counts matrix.}

\item{combos}{matrix; A matrix the combinations of cell types to be included
in the multiplets. Typically generated with the \code{combn} function.}

\item{adjustment;}{numeric A named numeric vector with length = the number of
unique cell types indicating the fraction of contribution for each cell
type. Names should indicate the corresponding cell type.}

\item{seed}{numeric; Seed for random number initiation. When generating many
multiplets by running the function multiple times in a loop, the seed should
be set dynamically in the function calling the makeSyntheticData function to
avoid all multiplets being identical.}

\item{...}{Additional arguments to pass on.}
}
\value{
A tibble with one gene per row and one synthesized multiplet per
 combo as columns.
}
\description{
Generates one multiplet per combo using the singlet counts data input to the
function.
}
\examples{

#use demo data


}
\author{
Jason T. Serviss
}
