% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{syntheticMultuplets}
\alias{syntheticMultuplets}
\title{syntheticMultuplets}
\arguments{
\item{nGenes}{numeric. Number of genes in generated synthetic data.}

\item{nCells}{numeric. Number of cells per cell type in generated synthetic
data.}

\item{nCellTypes}{numeric. Number of cell types in generated synthetic data.
For the frequency adjustment to work this should be an even number.}

\item{perplexity}{numeric. The perplexity parameter supplied to
spUnsupervised.}

\item{target}{numeric. The target percentage of connections for interacting
cells.}

\item{singletExpansion}{numeric. This argument is multiplied by the nCells
argument to give the total number of singlets synthesized per cell type.
nCells singlets are then reserved for the singlets used in testing whereas
the extra singlets created are used to adjust the connection frequencies
via the .adjustMultuplets and .adjustSelf units. Argument must be > 1.}

\item{...}{additional arguments to pass on}
}
\value{
A list containing singlets, multuplets, and the spUnsupervised
   object.
}
\description{
This unit synthesizes "all possible" multiplets combinations from the
singlets that were generated. Currently, doublets, triplets, and quadruplets
are synthesized. The unit also calls the adjustMultuplets and adjustSelf
units to adjust the connections in a manner that 1) self connections are the
most prevelant connection type (currently 2/3 connections) and 2) that each
cell type has a high percentage of connections to one other cell type.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
