% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticMultipletsFromCounts.R
\name{syntheticSinglets}
\alias{syntheticSinglets}
\title{syntheticSinglets}
\arguments{
\item{nGenes}{Number of genes in generated synthetic data.}

\item{nCells}{Number of cells per cell type in generated synthetic data.}

\item{nCellTypes}{Number of cell types in generated synthetic data.}

\item{...}{additional arguments to pass on}
}
\value{
A matrix with synthetic counts.
}
\description{
This unit uses the negative binomial distribution to synthesize the
singlets.
}
\examples{

synth <- syntheticSinglets(10, 10, 10)

}
\author{
Jason T. Serviss
}
\keyword{internal}
