% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{syntheticTestData}
\alias{syntheticTestData}
\title{syntheticTestData}
\arguments{
\item{nMultiplets}{Number of multiplets.}

\item{nGenes}{Number of genes in generated synthetic data.}

\item{nCells}{Number of cells per cell type in generated synthetic data.}

\item{nCellTypes}{Number of cell types in generated synthetic data. For the
frequency adjustment to work this should be an even number.}

\item{perplexity}{The perplexity parameter supplied to spUnsupervised.}

\item{target}{The target percentage of connections for interacting cells.}

\item{singletExpansion}{This argument is multiplied by the nCells argument to
give the total number of singlets synthesized per cell type. nCells
singlets are then reserved for the singlets used in testing whereas the
extra singlets created are used to adjust the connection frequencies via
the .adjustMultuplets and .adjustSelf units. Argument must be > 1.}

\item{...}{additional arguments to pass on}
}
\value{
A list containing variables syntheticData (all data generated),
   testSyntheticData (data to be used in the testing), uObj (spUnsupervised
   output using the generated singlets), table (a table of the connections
   synthesized).
}
\description{
This unit and the corresponding units synthesize the synthetic test data
that is subsequently saved and, as well, is used for testing.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
