% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingNewConditions.R
\name{testDistanceFunctions}
\alias{testDistanceFunctions}
\title{Tests multiple distance functions for comparison.}
\arguments{
\item{spCountsMul}{An spCounts object containing multiplets.}

\item{spUnsupervised}{An spUnsupervised object.}

\item{distFuns}{A named list of distance functions to test.}

\item{params}{A list with the same number of elements as distFuns containing
additional parameters that should be passed to each distFun.}

\item{...}{additional arguments to pass on}
}
\value{
The summary.
}
\description{
A wrapper for testing improvments in distance/cost functions.
}
\author{
Jason T. Serviss
}
