% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationFunctions.R
\name{tidyPermutationData}
\alias{tidyPermutationData}
\title{tidyPermutationData}
\arguments{
\item{sObj}{An spSwarm object where the "real" data is included.}

\item{permData}{Output from the permuteSwarm function.}

\item{...}{additional arguments to pass on}
}
\value{
A tibble with the real and permuted costs and fractions.
}
\description{
A function to wrangle the results from the \code{permuteSwarm} function.
}
\author{
Jason T. Serviss
}
