% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funToCheckKnownResults.R
\name{viewAsPlate}
\alias{viewAsPlate}
\title{viewAsPlate}
\arguments{
\item{plate}{Output from setup plate.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame
}
\description{
Takes the output from setupPlate and reformats it to a 96-well plate format.
}
\author{
Jason T. Serviss
}
