% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funToCheckKnownResults.R
\name{checkResults}
\alias{checkResults}
\title{checkResults}
\arguments{
\item{swarm}{CIMseqSwarm; A CIMseqSwarm object.}

\item{known}{A CIMseqSwarm object of the known result. The results should be
represented in the fractions slot according to the edge.cutoff provided.}

\item{edge.cutoff}{Standard definition.}

\item{...}{additional arguments to pass on}
}
\value{
data.frame
}
\description{
Untested for self connections.
}
\author{
Jason T. Serviss
}
