% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionDataTest.R
\name{expressionDataTest}
\alias{expressionDataTest}
\title{expressionDataTest}
\arguments{
\item{outPath}{Path to output results to.}

\item{cores}{Number of cores to be used for spSwarm.}

\item{n}{Number of multuplets.}

\item{ngenes}{Specifies the \code{n} argument to the \code{spTopMax}
function.}

\item{distFun}{The distance function to use during swarm optimization.}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
expressionDataTest()



}
\author{
Jason T. Serviss
}
\keyword{expressionDataTest}
