% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{syntheticDataTest}
\alias{syntheticDataTest}
\title{syntheticDataTest}
\arguments{
\item{outPath}{Path to output results to.}

\item{cores}{Number of cores to be used for spSwarm.}

\item{nMultiplets}{Number of multiplets.}

\item{nGenes}{Number of genes in generated synthetic data.}

\item{nCells}{Number of cells per cell type in generated synthetic data.}

\item{nCellTypes}{Number of cell types in generated synthetic data. For the
frequency adjustment to work this should be an even number.}

\item{distFun}{The distance function to use for swarm optimization.}

\item{target}{The target percentage of connections for interacting cells.}

\item{singletExpansion}{This argument is multiplied by the nCells argument to
give the total number of singlets synthesized per cell type. nCells
singlets are then reserved for the singlets used in testing whereas the
extra singlets created are used to adjust the connection frequencies via
the .adjustMultuplets and .adjustSelf units. Argument must be > 1.}

\item{perplexity}{The perplexity provided to Rtsne during the spUnsupervised
stage.}

\item{...}{additional arguments to pass on}
}
\value{
Saves results in .rda file.
}
\description{
Subtitle
}
\examples{

#use demo data
\dontrun{syntheticDataTest()}



}
\author{
Jason T. Serviss
}
