% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticDataTest.R
\name{syntheticSinglets}
\alias{syntheticSinglets}
\title{syntheticSinglets}
\arguments{
\item{nGenes}{Number of genes in generated synthetic data.}

\item{nCells}{Number of cells per cell type in generated synthetic data.}

\item{nCellTypes}{Number of cell types in generated synthetic data. For the
frequency adjustment to work this should be an even number.}

\item{singletExpansion}{This argument is multiplied by the nCells argument to
give the total number of singlets synthesized per cell type. nCells
singlets are then reserved for the singlets used in testing whereas the
extra singlets created are used to adjust the connection frequencies via
the .adjustMultuplets and .adjustSelf units. Argument must be > 1.}

\item{...}{additional arguments to pass on}
}
\value{
A list containing the singlets generated for connection adjustments
   and the singlets for use with spUnsupervised in the testing procedure.
}
\description{
This unit uses the negative binomial distribution to synthesize the
singlets used in the synthetic data test.
}
\examples{

cat("No example")

}
\author{
Jason T. Serviss
}
\keyword{internal}
